chore: remove `eqns` attributes that do not change equations (#17772)
As far as I can tell, the only use of `eqns` that we really need is in the `irreducible_def` command. Many other usages can be replaced with a few small changes, and a few do not seem to have been for older versions of Lean before they unfolded `match` in a nice way. This PR removes the usages of `eqns` that do not seem to change any unfolding behaviour in current versions of Lean.