chore: remove porting notes about adding/removing `@[ext]` attributes
All of these porting notes are about the change in behaviour of the `ext` tactic and attribute between Lean 3 and Lean 4: in Lean 4 it only accepts lemmas with conclusion `x = y` with `x` and `y` free variables, and only applies these if the types are reducibly equal. Since those behaviours are not likely to change in the future, we don't need to keep these porting notes.

I propose that we should keep #11041 since it indicates where these `ext` lemmas have not yet been added.

Closes: #5229
Closes: #11182