chore(Algebra/CharP): reduce theory requirements for `CharP.Defs` (#18182)
The `CharP/Defs.lean` file imports some structures and proves quite a bit of theory. This PR tries to ensure that the focus is on the definitions instead, namely `CharP`, `ringChar`, `ExpChar` and `ringExpChar`.

In the process, I made the following moves:
 * `CharP/Basic.lean`: renamed to `CharP/Lemmas.lean`
 * `CharP/Defs.lean`: all the lemmas that are not immediate consequences, or not needed to define `ringChar` or `expChar` have been moved to the new file `CharP/Basic.lean`

Note that `CharP/Basic.lean` and `CharP/Lemmas.lean` can be imported independently. They are both unstructured collections of results so I'm not sure how we want to clean that up further.