chore(Algebra/Group/Subgroup): split off `Defs` file
This PR splits off the definition of `Subgroup` and the `Group` instance into a new `Defs` file. Doing so required also moving the `Semigroup` instance on `Subsemigroup` and `Monoid` on `Submonoid` to the respective `Defs` file. So we needed to add one extra import: `Algebra.Group.InjSurj`. The imports and contents of that file are light enough in my opinion to make it worth it.