feat(NumberField/CanonicalEmbedding): define the `negAt` map (#18234)
Let `s` be a set of real places of the number field `K`. This PR defines the map `negAt s` on `mixedSpace K` that swaps the sign at all real places in `s` and proves some of its properties. 

It will be used later in #18231. 

This PR is part of the proof of the Analytic Class Number Formula.