chore: uniformize `ContDiffWithinAt` API with other local notions (#17927)
Currently, some statements are available for `DifferentiableWithinAt` but not for `ContinuousWithinAt` or conversely (and the same for the other `WithinAt` notions), or they are both available with slightly different names or assumptions.

This PR is a first step to uniformize these API: I've gone through the APIs of the different notions, and added all the lemmas that were available for another notion to `ContinuousWithinAt` if they were missing here, and fixed some names. Once this PR is merged, I will open other PRs to make sure that all the other notions also have the same full API.

Notable changes (imported from lemma names of other notions, and that I found better than the `ContinuousWithinAt` version):
* rename `ContinuousWithinAt.insert_self` to `ContinuousWithinAt.insert` 
* rename `continuousWithinAt_congr_nhds` to `continuousWithinAt_congr_set`, and change its assumptions from `𝓝[s] x = 𝓝[t] x` to `s =ᶠ[𝓝 x] t` (see Zulip discussion supporting this change at https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/local.20equality.20of.20sets/near/477653314)
* Added `continuousWithinAt_congr_set'` (taking an assumption `s =ᶠ[𝓝[{y}ᶜ] x] t`)
* Added several variations around `ContinuousWithinAt.congr_of_mem` (both in their implication and iff version)
* Added `of_eq` versions of all the composition lemmas

Also, use `variable` in the `ContinuousOn` file.