feat(NumberField/FundamentalCone): generalize the bijection to integral ideals (#18248)
The norm-preserving  bijection between the`integralPoint` in the fundamental cone and the integral ideals (times the torsion) is generalized to a bijection between the points of the fundamental cone lying in the image of an integral ideal `J` and the integral ideals divisible by `J` (times the torsion). 

This PR is part of the proof of the Analytic Class Number Formula.