RFC chore: do not import `Tactic.Common` in Mathlib
I went through the whole of Mathlib and replaced all imports of `Tactic.Common` with a more specialized file. I'm not sure if this is the right way to go, since it seems rather annoying to have to explicitly import basic tactics. But it should reduce some of the noise when testing how to split imports.