feat(NumberTheory/Padics): Mahler basis functions (#18335)
This PR adds the Mahler basis functions, which are the functions `x ↦ x.choose k` on `ℤ_[p]`. This is preparation for Mahler's theorem (to be added in a subsequent PR), which shows that these functions are a Banach space basis for the space of continuous functions on `ℤ_[p]`.

Partially based on the ETH bachelor thesis of Giulio Caflisch.