refactor: require that the range of a model with corners is included in the closure of its interior (#18403)
This will ensure that, for C^2 maps over real or complex manifolds, the second derivative in charts is symmetric (because that is true at interior points, and then it extends by continuity if interior points are dense). To be used in #18396 to define the Lie bracket of vector fields on manifolds.