chore: deprecate "goal-style" `tfae_have` syntax (#17045)
This PR deprecates the style of `tfae_have` syntax that behaves like Mathlib `have`, where the asserted hypothesis is prepended to the front of the goal list, in favor of `tfae_have ... := ...` syntax. It does so by preserving the old syntax and implementation in the same imports and logging a deprecation-styled warning. This warning can be turned off with `set_option Mathlib.Tactic.TFAE.useDeprecated true`.

This PR also includes some minor tweaks to `tfae_have` documentation and hover behavior to solidify `tfae_have ... := ...` as the current syntax.

It also modifies some old-style syntax that had crept into Mathlib in between the syntax removal (#11000) and this PR.