feat: quotient lift on a finite family (#5576)
In the future it may be used to prove that `ZFSet` is a model of ZF set theory without using `Classical.choice`. Specifically, it will be used to prove that all first-order formulas of `ZFSet` are definable.

Comparing to `(finChoice q).liftOn f`, `finLiftOn q f` asks users to provide a proof of `(∀ i, a i ≈ b i) → f a = f b` instead of `a ≈ b → f a = f b`, which is more readable and easier to rewrite. It also makes it easier for users to discover and use relevant APIs.



Co-authored-by: negiizhao <egresf@gmail.com>