feat: combinatorial part of the Faa di Bruno formula (#17169)
The Faa di Bruno formula expresses `D^n (g o f)` in terms of the iterated derivatives of `g` and `f`. The hardest part of the formalization is probably just being able to express the formula. This PR deals with the combinatorial prerequisites for the proof of the formula, defining ordered finpartitions of `Fin n` and showing how an ordered finpartition of `Fin (n+1)` can be built out of an ordered finpartition of `Fin n`.