refactor(GroupTheory/Sylow): Golf `Sylow.ofCard` (#18598)
`Sylow.ofCard` constructs a Sylow subgroup from a subgroup of the appropriate cardinality. Currently this construction is late in the file as it relies on the fact that Sylow subgroups (defined as maximal p-subgroups) have the appropriate cardinality. But the construction can be golfed and moved much earlier by instead using `IsPGroup.toSylow` which constructs a Sylow subgroup from a p-subgroup of index indivisible by p.