feat(Analysis/Normed/Ultra): Normed algebra preserves ultrametricity (#18698)
We show the result "Let `K` be a normed field. If a normed division ring `L` is a normed `K`-algebra, then `L` is ultrametric (i.e. the norm on `L` is nonarchimedean) if and only if `K` is." We also add the lemma of `IsNonarchimedean` is equivalent to `IsUltrametricDist`.