feat(NumberField/CanonicalEmbedding): put an euclidean space structure on the mixed space (#18245)
This PR defines the natural euclidian space stucture on the mixed space `ℝ^r₁ × ℂ^r₂` of a number field of signature of `K` and the map between the `Euclidean.mixedSpace K` and `mixedSpace`. Proves many properties of this map. In particular, it is volume preserving. 


This PR is part of the proof of the Analytic Class Number Formula.



Co-authored-by: Xavier Roblot <46200072+xroblot@users.noreply.github.com>