chore(RingTheory): reverse the Noetherian->Nilpotent import direction
This PR makes it so that `RingTheory/Nilpotent/Lemmas.lean` imports `RingTheory/Noetherian.lean` instead of the other way around. Not only does that better reflect the fact that `Nilpotent/Lemmas.lean` is a bit of a mix of random lemmas, this change should also clean up the import graph for `LinearAlgebra` somewhat, especially in combination with PR #18731.