chore: move GroupTheory.FiniteAbelian to GroupTheory.FiniteAbelian.Basic (#18729)
In preparation of adding material, this creates a new folder `Mathlib/GroupTheroy/FiniteAbelian` and moves `GroupTheory.FiniteAbelian` to `GroupTheory.FiniteAbelian.Basic`. This is (as of now) a leaf file, so the move has no further downstream effects. The new material will need more imports, so is better added in a separate file.