feat(CategoryTheory/Generator): introduce classes HasDetector and HasSeparator (#18577)
This change defines the typeclasses `HasDetector` and `HasSeparator` for categories with detectors/separators. Several results for `IsDetector` and `IsSeparator` are also proved for the new typeclasses.

Convenience theorems for proving the well-poweredness of categories with separators are added, too.

Two typos are fixed: "separating" is corrected to "coseparating" in the module docstring and `IsCospearator.isCodetector` is renamed to `IsCoseparator.isCodetector`.