fix: multigoal linter should not flag tactics that do nothing
When a tactic does not change the list of goals, the multigoal linter was giving incorrect error messages such as
```
There are 2 unclosed goals before 'change ...' and at least one remaining goal afterwards.
Please focus on the current goal, for instance using `·` (typed as "\.").
```
This PR adjusts the condition to allow all tactics that operate on one goal and result in zero or more goals. It also makes the error message more precise, for example
```
warning: The following tactic starts with 2 goals and ends with 1 goal, 1 of which is not operated on.
  exact .intro
Please focus on the current goal, for instance using `·` (typed as "\.").
note: this linter can be disabled with `set_option linter.style.multiGoal false`
```