feat: Goursat's lemma for subgroups Proves Goursat's lemma for subgroups.
If `I` is a subgroup of `G × H` which projects fully on both factors, then there exist normal
subgroups `G' ≤ G` and `H' ≤ H` such that the image of `I` in `G ⧸ G' × H ⧸ H'` is the graph of an
isomorphism `G ⧸ G' ≃ H ⧸ H'`.

We construct `G'` and `H'` as explicit subgroups.