feat: maintainer delegate everywhere (#18736)
The workflow file in this PR is meant to replace the three files for the maintainer merge from comment, review and review comment.

Besides having the same `maintainer merge` functionality, it also adds the `maintainer delegate` functionality.

I did not delete the old workflow files in the PR, since maybe, in order to properly test it, it may be a good idea to have the old and new workflows run in parallel for a final test.

In case you want to see the tests for the script, [this PR](https://github.com/bryangingechen/mathlib4/pull/14) has a live version of the workflow and you can test putting `maintainer merge/delegate` comments directly, from review or from review comment.

Note that the *test* workflow does *not* post to Zulip, nor does it check that who uses the maintainer merge feature is a reviewer.  Since the action does not post on Zulip, it posts the message instead in a PR comment.