fix(Algebra/Order/Field): fix instance diamond for `NNRat`
The `Semifield ℚ≥0` instance used a default definition of `zpow` using `zpowRec`, while the `LinearOrderedSemifield` instance has a custom definition of `zpow` called `Nonneg.zpow` (using the fact that `zpow` preserves nonnegativity). These two are not defeq and therefore we found an instance diamond in #18830.

The fix is to use `Nonneg.zpow` (or something defeq to it) in the definition of `NNRat.instSemifield`. However, this is tricky because `Nonneg.zpow` is only defined after that instance.

I saw two ways to implement the fix:
 * postpone the `Semifield NNRat` instance to after the definition of `Nonneg.zpow`. It should definitely be possible to move all the `NNRat`-related definitions from `Algebra.Field.Rat` to `Algebra.Order.Field.Rat`, but it looks like the import graph gets a bit more tangled this way.
 * redo the proof of `Nonneg.zpow` inline in `NNRat.instSemifield`. It turns out `inv` and `div` are already redefined in the same way in the same section, so this seems like the neatest solution with the least surprises.

I also added an `example` that checks the diamond is indeed now fixed.