feat(Analysis/Complex/AbsMax): add two lemmas (#18198)
add two results:

-  If `f` is differentiable on the open unit ball `{z : ℂ | ‖z‖ < 1}`, and `‖f‖` attains a maximum
in  this open ball, then `f` is constant 

-   If `f` is a function differentiable on the open unit ball, and there exists an `r < 1` such that
any value of `‖f‖` on the open ball is bounded above by some value on the closed ball of radius `r`,
then `f` is constant. 

These lemmas together with #18192 will be used to prove some dimension results for spaces of modular forms.