feat: tensoring a localization with itself does nothing (#18744)
`IsLocalization.tensorSelfAlgEquiv`: If A is a localization of a commutative ring R, then `A ⊗[R] A` is canonically isomorphic to `A` as `A`-algebras via multiplication. To construct the isomorphism we introduce a stronger version of `Algebra.TensorProduct.lmul'`.

Also add some easy theorem (`linearMap_compatibleSMul`) and defs (`distribBaseChange`, `LinearEquiv.baseChange`) and protect `IsLocalization.map_smul`.