fix(Tactic/Linter): the `minImports` linter understands file edits
Previously, the `minImports` linter stored the state of the last command to be elaborated. This does not work if we edit a file, so we should instead store state per command, and look up the correct command to continue from. Since editing a file will restart elaboration at the command that started *before* the edit, if we track the *ending* position of a command, we know which commands are before and after the edit.

I tested the changes manually (and tweaked the exact edit detection algorithm) with many edits to `MathlibTests/MinImports.lean` but I wouldn't know an automated way of doing tests here, unfortunately.