chore(LinearAlgebra/Matrix): split off `Matrix.stdBasis` to its own file (#18733)
Previously, `LinearAlgebra.StdBasis` also defined the standard basis for matrices. We don't need matrices to define finite-dimensional spaces.

It would be nice to instead swap the import order: `LinearAlgebra.StdBasis` imported by `LinearAlgebra.Matrix.Basis`, but unfortunately that would cause a cycle with `LinearAlgebra.Matrix.ToLin`. Not entirely sure if that dependency is worth the hassle of further untangling.