feat: weighted technical debt in PR summary (#19020)
Add a "weighted technical debt average" to the PR summary.

The weighted average is obtained by weighting each change by the inverse of the current value of the measure.  Thus, to balance an increase in a category that has a small number of exceptions, you would have to decrease substantially more a category that has a large number of exceptions.

The heuristic behind this choice is that category with few exceptions tend to be harder/more desirable to remove.  This is not perfect, but it is a start!